requirements.txt
aiogram>=3.4
aiosqlite>=0.20
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
